Working iteration with the example text given.
This is the first working iteration of the assignment with the text example text that was given.

Not checking sentence indexes yet, this is only that the same amount of words are picked up and that a couple of the word counts match between the example and my implementation.

My implementation uses a buffered reader to read the file into a string. This assumes that whatever is running this program will have enough memory for that. It then passes that to a BreakIterator, which is something I didn't even know existed, to break it apart into sentences. Then each sentence is broken into words and each word is added into the results or if it is already present, the count is updated and the sentence index is added if it doesn't exist already.