Final Product
Adjusted the sentence count to not check for duplicates and just add it everytime it finds it. It looks like the bonus sentence might have been rewritten based on the output and what the sentence says.

Added in the code so that the main method prints everything out in a nice table in alphabetical order. Stole a function off stackoverflow to convert the index into a letter. Also stole some fancy new Java streams stuff to get everything in alphabetical order. Used string format with padding to get the whitespaces to line up.

You should be able to drop a different txt file into the `src/main/resources` to try out a different set of text. Then just type it into the prompt.